requirements_sprint_2_project.txt
torch>=2.2,<3
transformers>=4.38,<5
datasets>=2.18,<3
evaluate>=0.4,<0.5
rouge_score>=0.1.2,<0.2
numpy>=1.26,<3
pandas>=2.2,<3
scikit-learn>=1.3,<2
tqdm>=4.65,<5
PyYAML>=6.0,<7
jupyter>=1.0,<2
ipykernel>=6,<7
nltk>=3.9,<4

